src/api/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.7.0
cryptography==42.0.1
firebase-admin==6.4.0
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
msgpack==1.0.7
multidict==6.0.5
mysql-connector-python==8.3.0
numpy==1.26.3
pandas==2.2.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
